requirements.txt
Django==4.2.25
djangorestframework==3.14.0
pymongo==4.5.0
python-dotenv==1.0.0
Pillow==10.0.1
django-cors-headers==4.3.1
dnspython==2.4.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0